라이브 시청자수 리턴

+++ backend/src/main/java/com/ssafy/api/controller/LiveController.java
@@ -85,7 +85,7 @@ public class LiveController {
 	@Transactional
 	@PostMapping(value="/live/live-viewer")
 	public ResponseEntity updateViewerCount(@RequestBody LiveVIewerReq liveViewerReq){
-		long a  = liveService.updateViewerCount(liveViewerReq);
+		int a  = liveService.updateViewerCount(liveViewerReq);
 		return new ResponseEntity(a,HttpStatus.OK);
 	}
